inventory stuff
player access inventory by typing inventory (from main.cpp) which brings player to inventory menu (playersInventory.cpp)
the player than then use (countInventory.cpp) to list the amount of items in his inventory, or
the player can use (listInventory.cpp) to list the items in his inventory. (still working on it. atm all it does is cout text to show it runs)
I also have an if statement for future use of items (use, examine) in the (playersInventory.cpp)